fix eval freq to float

+++ configs.py
@@ -73,7 +73,7 @@ def parse_arguments():
     parser.add_argument('--task_num', default=5, type=int)
     parser.add_argument('--early_stop', action='store_true')
     parser.add_argument('--patience', type=int, default=2)
-    parser.add_argument('--eval_freq', type=int, default=5)
+    parser.add_argument('--eval_freq', type=float, default=5)
 
     parser.add_argument('--input_map', action='store_true', help="Whether to use input mapping, if False, use span_s to predict trigger type")
     parser.add_argument('--class_num', type=int, default=10)
